scripts/test_live_chaos.py
"""
HydroTwin OS — Chaos Monkey Integration Test

Simulates an adversarial infrastructure environment:
1. Verifies baseline observability and distributed mesh integrity.
2. Injects a catastrophic failure by halting the live Kafka KRaft container.
3. Validates that the FastAPI inference router does not crash (failsafe/stale cache mechanism).
4. Restores Kafka and verifies that the `run_live_datacenter` node and RL worker immediately reconnect and resume telemetry flow without deadlock.
"""

import time
import subprocess
import logging
import requests

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger("ChaosMonkey")

def run_chaos_test():
    logger.info("Starting Chaos Monkey on Live Event Mesh...")
    
    # 1. Baseline checkpoint
    try:
        res = requests.get("http://localhost:8000/api/metrics", timeout=5)
        data = res.json()
        logger.info(f"Baseline ✅ API alive. Step: {data['step']}, Inlet Temp: {data['physics']['inlet_temp_c']}°C")
    except Exception as e:
        logger.error(f"FastAPI backend unresponsive: {e}")
        return
        
    # 2. Inject Failure
    logger.warning("💀 INJECTING SERVICE FAILURE: Halting Kafka KRaft Broker...")
    subprocess.run(["docker", "stop", "hydrotwin-kafka"], check=True)
    
    logger.info("Kafka node offline. Waiting 15 seconds to simulate outage...")
    time.sleep(15)
    
    # 3. Assert Backend Survival (Failsafe cache)
    try:
        res = requests.get("http://localhost:8000/api/metrics", timeout=5)
        logger.info("Outage Check ✅ FastAPI backend survived Kafka outage without memory deadlock.")
    except Exception as e:
        logger.error(f"❌ FastAPI CRASHED during outage: {e}")
        
    # 4. Restore Infrastructure
    logger.info("🔧 RESTORING INFRASTRUCTURE: Starting Kafka KRaft Broker...")
    subprocess.run(["docker", "start", "hydrotwin-kafka"], check=True)
    
    logger.info("Kafka node online. Waiting 20 seconds for cluster partition reelection and client reconnections...")
    time.sleep(20)
    
    # 5. Assert Recovery
    try:
        res = requests.get("http://localhost:8000/api/metrics", timeout=5)
        data = res.json()
        logger.info(f"Recovery Check ✅ System re-synchronized. Latest RL Policy Action: {data['rl_decision']['action']}")
        logger.info(f"Final Inlet Temp reading: {data['physics']['inlet_temp_c']}°C")
    except Exception as e:
        logger.error(f"❌ System failed to re-synchronize: {e}")
        return
        
    logger.info("🏆 Chaos Test Complete! HydroTwin OS is provably fault-tolerant.")

if __name__ == "__main__":
    run_chaos_test()
