pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hydrotwin"
version = "0.1.0"
description = "HydroTwin OS — Plane 3: Water-Energy-Carbon Nexus Agent"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Proprietary"}
authors = [
    {name = "HydroTwin Team"}
]

dependencies = [
    # RL & Deep Learning
    "stable-baselines3>=2.3.0",
    "torch>=2.1.0",
    "gymnasium>=0.29.0",
    "tensorboard>=2.15.0",

    # Forecasting
    "prophet>=1.1.5",

    # Data & Science
    "numpy>=1.24.0",
    "pandas>=2.1.0",
    "scipy>=1.11.0",
    "matplotlib>=3.8.0",

    # API & Web
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.25.0",
    "httpx>=0.26.0",
    "websockets>=12.0",

    # Messaging
    "kafka-python>=2.0.2",

    # Configuration & Serialization
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",

    # Database Clients
    "influxdb-client>=1.38.0",
    "neo4j>=5.15.0",

    # Observability
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]

[tool.setuptools.packages.find]
include = ["hydrotwin*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py310"
